support homebrew on macOS + update ci + new debug logger

+++ internal/ci/linux.sh
@@ -17,12 +17,20 @@ sudo apt-get -y -qq install gcc-5 g++-5
 sudo rm /usr/bin/gcc; sudo ln -s /usr/bin/gcc-5 /usr/bin/gcc
 sudo rm /usr/bin/g++; sudo ln -s /usr/bin/g++-5 /usr/bin/g++
 
-#download and install qt
-QT=qt-opensource-linux-x64-android-5.7.0.run
-curl -sL --retry 10 --retry-delay 10 -o /tmp/$QT https://download.qt.io/official_releases/qt/5.7/5.7.0/$QT
-chmod +x /tmp/$QT
-/tmp/$QT --script $GOPATH/src/github.com/therecipe/qt/internal/ci/iscript.qs
-rm -f /tmp/$QT
+if [ "$QT_PKG_CONFIG" == "true" ]
+then
+  #download and install qt
+  sudo add-apt-repository -y ppa:beineri/opt-qt57-trusty
+  sudo apt-get -qq update
+  sudo apt-get -y -qq install qt57*
+else
+  #download and install qt
+  QT=qt-opensource-linux-x64-android-5.7.0.run
+  curl -sL --retry 10 --retry-delay 10 -o /tmp/$QT https://download.qt.io/official_releases/qt/5.7/5.7.0/$QT
+  chmod +x /tmp/$QT
+  /tmp/$QT --script $GOPATH/src/github.com/therecipe/qt/internal/ci/iscript.qs
+  rm -f /tmp/$QT
+fi
 
 #download and install android sdk
 SDK=android-sdk_r24.4.1-linux.tgz
